(BE) {rename} app.js에서 facilities.js 분리
