+++ server/app.js
@@ -1,6 +1,5 @@
 require('dotenv').config();
 const express = require("express");
-const mysql = require("./mysql");
 const nodePath = require("path");
 const cors = require("cors");
 
@@ -22,45 +21,13 @@ app.get("/search", (req, res) => {
   res.sendFile(nodePath.join(__dirname, "public", "search.html"));
 });
 
-// 지도에 위치 표시 
-app.get("/facilities", (req, res) => {
-  const { type, keyword } = req.query;
-  let query = "SELECT * FROM medical_facilities WHERE 1=1";
-  const values = [];
-
-  if (type) {
-    query += " AND type = ?";
-    values.push(type);
-  }
-
-  if (keyword) {
-    query += " AND (bplcnm LIKE ? OR rdnwhladdr LIKE ? OR sitewhladdr LIKE ?)";
-    values.push(`%${keyword}%`, `%${keyword}%`, `%${keyword}%`);
-  }
-
-  query += " LIMIT 10000";
-
-  console.log("Executing query:", query);
-  console.log("Query values:", values);
-
-  mysql.query(query, values, (err, results) => {
-    if (err) {
-      console.error('Database query error:', err);
-      console.error('Error details:', JSON.stringify(err, null, 2));
-      return res.status(500).json({ error: 'Internal server error', details: err.message });
-    }
-
-    console.log(`Query returned ${results.length} results`);
-    res.json(results);
-  });
-});
-
 // 라우터 설정
 const categoryRouter = require('./routes/category');
 const postRouter = require('./routes/post');
 const reviewRouter = require('./routes/review');
 const commentRouter = require('./routes/comment');
 const authRouter = require('./routes/auth');
+const facilitiesRouter = require('./routes/facilities');  // 새로 추가
 const mypageRouter = require('./routes/mypage')
 
 app.use('/category', categoryRouter);
@@ -68,6 +35,7 @@ app.use('/posts', postRouter);
 app.use('/reviews', reviewRouter);
 app.use('/comments', commentRouter);
 app.use('/auth', authRouter);
+app.use('/facilities', facilitiesRouter);  // 새로 추가
 app.use('/mypage', mypageRouter);
 
 // 서버 시작
